TST: update unit tests
